// client/src/util/actions.ts
export const enum ActionTypes {
  SET_TOKEN = "SET_TOKEN",
  SET_USER = "SET_USER",
  SET_PLAYLISTS = "SET_PLAYLISTS",
  SET_PLAYING = "SET_PLAYING",
  SET_PLAYER_STATE = "SET_PLAYER_STATE",
  SET_SHUFFLE = "SET_SHUFFLE",
  SET_PLAYLIST = "SET_PLAYLIST",
  SET_PLAYLIST_ID = "SET_PLAYLIST_ID",
  SET_REPEAT = "SET_REPEAT",
}